Added generic shutdown method
Works only for RetroPie!
Please copy ./multi_switch.sh

+++ SafeShutdown.py
@@ -2,6 +2,7 @@
 from gpiozero import Button, LED
 import os 
 from signal import pause
+import subprocess
 
 powerPin = 3 
 resetPin = 2 
@@ -16,15 +17,25 @@ power.on()
 #functions that handle button events
 def when_pressed():
  led.blink(.2,.2)
- os.system("sudo shutdown -h now")
+ output = subprocess.check_output(['./multi_switch.sh', '--es-pid']).strip()
+ if output != "0":
+     os.system("./multi_switch.sh --es-poweroff")
+ else:
+     os.system("sudo shutdown -h now")
+     
 def when_released():
  led.on()
 def reboot():
- os.system("sudo reboot")
-  
+ output = subprocess.check_output(['./multi_switch.sh', '--es-pid']).strip()
+ if output != "0":
+     os.system("./multi_switch.sh --es-restart")
+ else:
+     os.system("sudo reboot")
+ 
 btn = Button(powerPin, hold_time=hold)
 rebootBtn = Button(resetPin)
 rebootBtn.when_pressed = reboot 
 btn.when_pressed = when_pressed
 btn.when_released = when_released
 pause()
+
